tests: more test output improving
