+++ tests/outbound/index.js
@@ -3,7 +3,6 @@
 const assert = require('node:assert')
 const fs   = require('fs');
 const path = require('path');
-const os   = require('os');
 
 const constants = require('haraka-constants');
 const logger = require('../../logger');
@@ -18,6 +17,7 @@ const lines = [
 ];
 
 describe('outbound', () => {
+
     it('converts \\n and \\r\\n line endings to \\r\\n' , () => {
 
         for (const ending of ['\n', '\r\n']) {
@@ -30,7 +30,6 @@ describe('outbound', () => {
             while ((match = re.exec(contents))) {
                 let line = match[1];
                 line = line.replace(/\r?\n?$/, '\r\n'); // assure \r\n ending
-                // transaction.add_data(new Buffer(line));
                 result += line;
                 contents = contents.substr(match[1].length);
                 if (contents.length === 0) {
@@ -74,227 +73,148 @@ describe('outbound', () => {
         config.set_temp_fail_intervals();
         assert.deepEqual(config.cfg.temp_fail_intervals, [64,128,256,512,1024,2048,4096,8192,16384,32768,65536,131072]);
     })
-})
 
-describe('qfile', () => {
-    beforeEach((done) => {
-        this.qfile = require('../../outbound').qfile;
-        done();
-    })
+    describe('get_tls_options', () => {
+        beforeEach((done) => {
+            process.env.HARAKA_TEST_DIR=path.resolve('tests');
+            this.outbound = require('../../outbound');
+            this.obtls = require('../../outbound/tls');
+            const tls_socket = require('../../tls_socket');
 
-    it('name() basic functions', () => {
-        const name = this.qfile.name();
-        const split = name.split('_');
-        assert.equal(split.length, 7);
-        assert.equal(split[2], 0);
-        assert.equal(split[3], process.pid);
-    })
-
-    it('name() with overrides', () => {
-        const overrides = {
-            arrival : 12345,
-            next_attempt : 12345,
-            attempts : 15,
-            pid : process.pid,
-            uid : 'XXYYZZ',
-            host : os.hostname(),
-        };
-        const name = this.qfile.name(overrides);
-        const split = name.split('_');
-        assert.equal(split.length, 7);
-        assert.equal(split[0], overrides.arrival);
-        assert.equal(split[1], overrides.next_attempt);
-        assert.equal(split[2], overrides.attempts);
-        assert.equal(split[3], overrides.pid);
-        assert.equal(split[4], overrides.uid);
-        assert.equal(split[6], overrides.host);
-    })
-
-    it('rnd_unique() is unique-ish', () => {
-        const repeats = 1000;
-        const u = this.qfile.rnd_unique();
-        for (let i = 0; i < repeats; i++){
-            assert.notEqual(u, this.qfile.rnd_unique());
-        }
-    })
-
-    it('parts() updates previous queue filenames', () => {
-        // $nextattempt_$attempts_$pid_$uniq.$host
-        const name = "1111_0_2222_3333.foo.example.com"
-        const parts = this.qfile.parts(name);
-        assert.equal(parts.next_attempt, 1111);
-        assert.equal(parts.attempts, 0);
-        assert.equal(parts.pid, 2222);
-        assert.equal(parts.host, 'foo.example.com');
-    })
-
-    it('parts() handles standard queue filenames', () => {
-        const overrides = {
-            arrival : 12345,
-            next_attempt : 12345,
-            attempts : 15,
-            pid : process.pid,
-            uid : 'XXYYZZ',
-            host : os.hostname(),
-        };
-        const name = this.qfile.name(overrides);
-        const parts = this.qfile.parts(name);
-        assert.equal(parts.arrival, overrides.arrival);
-        assert.equal(parts.next_attempt, overrides.next_attempt);
-        assert.equal(parts.attempts, overrides.attempts);
-        assert.equal(parts.pid, overrides.pid);
-        assert.equal(parts.uid, overrides.uid);
-        assert.equal(parts.host, overrides.host);
-    })
-})
-
-describe('get_tls_options', () => {
-    beforeEach((done) => {
-        process.env.HARAKA_TEST_DIR=path.resolve('tests');
-        this.outbound = require('../../outbound');
-        this.obtls = require('../../outbound/tls');
-        const tls_socket = require('../../tls_socket');
-
-        // reset config to load from tests directory
-        const testDir = path.resolve('tests');
-        this.outbound.config = this.outbound.config.module_config(testDir);
-        this.obtls.test_config(tls_socket.config.module_config(testDir), this.outbound.config);
-        this.obtls.init(done)
-    })
-
-    afterEach((done) => {
-        delete process.env.HARAKA_TEST_DIR;
-        done();
-    })
-
-    it('gets TLS properties from tls.ini.outbound', () => {
-        const tls_config = this.obtls.get_tls_options(
-            { exchange: 'mail.example.com'}
-        );
-
-        assert.deepEqual(tls_config, {
-            servername: 'mail.example.com',
-            key: fs.readFileSync(path.resolve('tests','config','outbound_tls_key.pem')),
-            cert: fs.readFileSync(path.resolve('tests','config','outbound_tls_cert.pem')),
-            dhparam: fs.readFileSync(path.resolve('tests','config','dhparams.pem')),
-            ciphers: 'ECDHE-RSA-AES256-GCM-SHA384',
-            minVersion: 'TLSv1',
-            rejectUnauthorized: false,
-            requestCert: false,
-            honorCipherOrder: false,
-            redis: { disable_for_failed_hosts: false },
-            no_tls_hosts: [],
-            force_tls_hosts: ['first.example.com', 'second.example.net']
+            // reset config to load from tests directory
+            const testDir = path.resolve('tests');
+            this.outbound.config = this.outbound.config.module_config(testDir);
+            this.obtls.test_config(tls_socket.config.module_config(testDir), this.outbound.config);
+            this.obtls.init(done)
         })
-    })
-})
 
-describe('build_todo', () => {
-    beforeEach((done) => {
-        this.outbound = require('../../outbound');
-        try {
-            fs.unlinkSync('tests/queue/multibyte');
-            fs.unlinkSync('tests/queue/plain');
-        }
-        catch (ignore) {}
-        done();
-    })
+        afterEach((done) => {
+            delete process.env.HARAKA_TEST_DIR;
+            done();
+        })
 
-    it('saves a file', () => {
-        const todo = JSON.parse('{"queue_time":1507509981169,"domain":"redacteed.com","rcpt_to":[{"original":"<postmaster@redacteed.com>","original_host":"redacteed.com","host":"redacteed.com","user":"postmaster"}],"mail_from":{"original":"<matt@tnpi.net>","original_host":"tnpi.net","host":"tnpi.net","user":"matt"},"notes":{"authentication_results":["spf=pass smtp.mailfrom=tnpi.net"],"spf_mail_result":"Pass","spf_mail_record":"v=spf1 a mx include:mx.theartfarm.com ?include:forwards._spf.tnpi.net include:lists._spf.tnpi.net -all","attachment_count":0,"attachments":[{"ctype":"application/pdf","filename":"FileWithoutAccent Chars.pdf","extension":".pdf","md5":"6c1d5f5c047cff3f6320b1210970bdf6"}],"attachment_ctypes":["application/pdf","multipart/mixed","text/plain","application/pdf"],"attachment_files":["FileWithoutaccent Chars.pdf"],"attachment_archive_files":[]},"uuid":"1D5483B0-3E00-4280-A961-3AFD2017B4FC.1"}');
-        const fd = fs.openSync('tests/queue/plain', 'w');
-        const ws = new fs.createWriteStream('tests/queue/plain', { fd, flags: constants.WRITE_EXCL });
-        ws.on('close', () => {
-            // console.log(arguments);
-            assert.ok(1);
-        })
-        ws.on('error', (e) => {
-            console.error(e);
-        })
-        this.outbound.build_todo(todo, ws, () => {
-            ws.write(Buffer.from('This is the message body'));
-            fs.fsync(fd, () => { ws.close(); })
+        it('gets TLS properties from tls.ini.outbound', () => {
+            const tls_config = this.obtls.get_tls_options(
+                { exchange: 'mail.example.com'}
+            );
+
+            assert.deepEqual(tls_config, {
+                servername: 'mail.example.com',
+                key: fs.readFileSync(path.resolve('tests','config','outbound_tls_key.pem')),
+                cert: fs.readFileSync(path.resolve('tests','config','outbound_tls_cert.pem')),
+                dhparam: fs.readFileSync(path.resolve('tests','config','dhparams.pem')),
+                ciphers: 'ECDHE-RSA-AES256-GCM-SHA384',
+                minVersion: 'TLSv1',
+                rejectUnauthorized: false,
+                requestCert: false,
+                honorCipherOrder: false,
+                redis: { disable_for_failed_hosts: false },
+                no_tls_hosts: [],
+                force_tls_hosts: ['first.example.com', 'second.example.net']
+            })
         })
     })
 
-    it('saves a file with multibyte chars', () => {
-        const todo = JSON.parse('{"queue_time":1507509981169,"domain":"redacteed.com","rcpt_to":[{"original":"<postmaster@redacteed.com>","original_host":"redacteed.com","host":"redacteed.com","user":"postmaster"}],"mail_from":{"original":"<matt@tnpi.net>","original_host":"tnpi.net","host":"tnpi.net","user":"matt"},"notes":{"authentication_results":["spf=pass smtp.mailfrom=tnpi.net"],"spf_mail_result":"Pass","spf_mail_record":"v=spf1 a mx include:mx.theartfarm.com ?include:forwards._spf.tnpi.net include:lists._spf.tnpi.net -all","attachment_count":0,"attachments":[{"ctype":"application/pdf","filename":"FileWîthÁccent Chars.pdf","extension":".pdf","md5":"6c1d5f5c047cff3f6320b1210970bdf6"}],"attachment_ctypes":["application/pdf","multipart/mixed","text/plain","application/pdf"],"attachment_files":["FileWîthÁccent Chars.pdf"],"attachment_archive_files":[]},"uuid":"1D5483B0-3E00-4280-A961-3AFD2017B4FC.1"}');
-        const fd = fs.openSync('tests/queue/multibyte', 'w');
-        const ws = new fs.WriteStream('tests/queue/multibyte', { fd, flags: constants.WRITE_EXCL });
-        ws.on('close', () => {
-            assert.ok(1);
+    describe('build_todo', () => {
+        beforeEach((done) => {
+            this.outbound = require('../../outbound');
+            try {
+                fs.unlinkSync('tests/queue/multibyte');
+                fs.unlinkSync('tests/queue/plain');
+            }
+            catch (ignore) {}
+            done();
         })
-        ws.on('error', (e) => {
-            console.error(e);
+
+        it('saves a file', () => {
+            const todo = JSON.parse('{"queue_time":1507509981169,"domain":"redacteed.com","rcpt_to":[{"original":"<postmaster@redacteed.com>","original_host":"redacteed.com","host":"redacteed.com","user":"postmaster"}],"mail_from":{"original":"<matt@tnpi.net>","original_host":"tnpi.net","host":"tnpi.net","user":"matt"},"notes":{"authentication_results":["spf=pass smtp.mailfrom=tnpi.net"],"spf_mail_result":"Pass","spf_mail_record":"v=spf1 a mx include:mx.theartfarm.com ?include:forwards._spf.tnpi.net include:lists._spf.tnpi.net -all","attachment_count":0,"attachments":[{"ctype":"application/pdf","filename":"FileWithoutAccent Chars.pdf","extension":".pdf","md5":"6c1d5f5c047cff3f6320b1210970bdf6"}],"attachment_ctypes":["application/pdf","multipart/mixed","text/plain","application/pdf"],"attachment_files":["FileWithoutaccent Chars.pdf"],"attachment_archive_files":[]},"uuid":"1D5483B0-3E00-4280-A961-3AFD2017B4FC.1"}');
+            const fd = fs.openSync('tests/queue/plain', 'w');
+            const ws = new fs.createWriteStream('tests/queue/plain', { fd, flags: constants.WRITE_EXCL });
+            ws.on('close', () => {
+                // console.log(arguments);
+                assert.ok(1);
+            })
+            ws.on('error', (e) => {
+                console.error(e);
+            })
+            this.outbound.build_todo(todo, ws, () => {
+                ws.write(Buffer.from('This is the message body'));
+                fs.fsync(fd, () => { ws.close(); })
+            })
         })
-        this.outbound.build_todo(todo, ws, () => {
-            ws.write(Buffer.from('This is the message body'));
-            fs.fsync(fd, () => { ws.close(); })
+
+        it('saves a file with multibyte chars', () => {
+            const todo = JSON.parse('{"queue_time":1507509981169,"domain":"redacteed.com","rcpt_to":[{"original":"<postmaster@redacteed.com>","original_host":"redacteed.com","host":"redacteed.com","user":"postmaster"}],"mail_from":{"original":"<matt@tnpi.net>","original_host":"tnpi.net","host":"tnpi.net","user":"matt"},"notes":{"authentication_results":["spf=pass smtp.mailfrom=tnpi.net"],"spf_mail_result":"Pass","spf_mail_record":"v=spf1 a mx include:mx.theartfarm.com ?include:forwards._spf.tnpi.net include:lists._spf.tnpi.net -all","attachment_count":0,"attachments":[{"ctype":"application/pdf","filename":"FileWîthÁccent Chars.pdf","extension":".pdf","md5":"6c1d5f5c047cff3f6320b1210970bdf6"}],"attachment_ctypes":["application/pdf","multipart/mixed","text/plain","application/pdf"],"attachment_files":["FileWîthÁccent Chars.pdf"],"attachment_archive_files":[]},"uuid":"1D5483B0-3E00-4280-A961-3AFD2017B4FC.1"}');
+            const fd = fs.openSync('tests/queue/multibyte', 'w');
+            const ws = new fs.WriteStream('tests/queue/multibyte', { fd, flags: constants.WRITE_EXCL });
+            ws.on('close', () => {
+                assert.ok(1);
+            })
+            ws.on('error', (e) => {
+                console.error(e);
+            })
+            this.outbound.build_todo(todo, ws, () => {
+                ws.write(Buffer.from('This is the message body'));
+                fs.fsync(fd, () => { ws.close(); })
+            })
         })
     })
-})
 
-describe('timer_queue', () => {
-    beforeEach((done) => {
-        process.env.HARAKA_TEST_DIR=path.resolve('tests');
-        this.outbound = require('../../outbound');
-        const TimerQueue = require('../../outbound/timer_queue');
-        this.ob_timer_queue = new TimerQueue(500);
-        done();
-    })
+    describe('timer_queue', () => {
+        beforeEach((done) => {
+            process.env.HARAKA_TEST_DIR=path.resolve('tests');
+            this.outbound = require('../../outbound');
+            const TimerQueue = require('../../outbound/timer_queue');
+            this.ob_timer_queue = new TimerQueue(500);
+            done();
+        })
 
-    afterEach((done) => {
-        delete process.env.HARAKA_TEST_DIR;
-        this.ob_timer_queue.shutdown();
-        done()
-    })
+        afterEach((done) => {
+            delete process.env.HARAKA_TEST_DIR;
+            this.ob_timer_queue.shutdown();
+            done()
+        })
 
-    it('has initial length of 0', () => {
+        it('has initial length of 0', () => {
+            assert.equal(this.ob_timer_queue.length(), 0);
+        })
 
-        const tq_length = this.ob_timer_queue.length();
+        it('can add items', () => {
+            this.ob_timer_queue.add("1", 1000);
+            this.ob_timer_queue.add("2", 2000);
 
-        assert.equal(tq_length, 0);
-    })
+            assert.equal(this.ob_timer_queue.length(), 2);
+        })
 
-    it('can add items', () => {
+        it('can drain items', () => {
 
-        this.ob_timer_queue.add("1", 1000);
-        this.ob_timer_queue.add("2", 2000);
+            this.ob_timer_queue.add("1", 1000);
+            this.ob_timer_queue.add("2", 2000);
 
-        const tq_length = this.ob_timer_queue.length();
+            let tq_length = this.ob_timer_queue.length();
 
-        assert.equal(tq_length, 2);
-    })
+            assert.equal(tq_length, 2);
 
-    it('can drain items', () => {
+            this.ob_timer_queue.drain();
+            tq_length = this.ob_timer_queue.length();
 
-        this.ob_timer_queue.add("1", 1000);
-        this.ob_timer_queue.add("2", 2000);
+            assert.equal(tq_length, 0);
+        })
 
-        let tq_length = this.ob_timer_queue.length();
+        it('can discard items by id', () => {
 
-        assert.equal(tq_length, 2);
+            this.ob_timer_queue.add("1", 1000);
+            this.ob_timer_queue.add("2", 2000);
 
-        this.ob_timer_queue.drain();
-        tq_length = this.ob_timer_queue.length();
+            let tq_length = this.ob_timer_queue.length();
 
-        assert.equal(tq_length, 0);
-    })
+            assert.equal(tq_length, 2);
 
-    it('can discard items by id', () => {
+            this.ob_timer_queue.discard("2");
+            tq_length = this.ob_timer_queue.length();
 
-        this.ob_timer_queue.add("1", 1000);
-        this.ob_timer_queue.add("2", 2000);
-
-        let tq_length = this.ob_timer_queue.length();
-
-        assert.equal(tq_length, 2);
-
-        this.ob_timer_queue.discard("2");
-        tq_length = this.ob_timer_queue.length();
-
-        assert.equal(tq_length, 1);
-        assert.equal(this.ob_timer_queue.queue[0].id, "1");
-
+            assert.equal(tq_length, 1);
+            assert.equal(this.ob_timer_queue.queue[0].id, "1");
+        })
     })
 })
